30 - selectors and slice with unit tests
