+++ src/entities/counter/model/selectors/counterSelectors.ts
