+++ src/entities/counter/model/slice/counterSlice.ts
