+++ src/entities/counter/model/slice/index.ts
@@ -1,0 +1,1 @@
+export { counterActions, counterReducer } from './counterSlice';
