+++ src/entities/counter/ui/index.ts
@@ -1,0 +1,1 @@
+export { Counter } from './counter';
